+++ resources/assets/js/admin/form/file.js
@@ -55,17 +55,9 @@ Vue.component('element-file', Vue.extend({
         remove () {
             var self = this;
 
-            swal({
-                title: i18next.t('lang.message.are_you_sure'),
-                type: 'warning',
-                showCancelButton: true,
-                confirmButtonColor: '#3085d6',
-                cancelButtonColor: '#d33',
-                confirmButtonText: i18next.t('lang.button.yes')
-            }).then(() => {
+            Admin.Messages.confirm(trans('lang.message.are_you_sure')).then(() => {
                 self.value = '';
-            }, dismiss => {
-
+                self.file = '';
             });
         },
         closeAlert () {
@@ -75,6 +67,9 @@ Vue.component('element-file', Vue.extend({
     computed: {
         has_value () {
             return this.value.length > 0
+        },
+        file () {
+            return ((this.value.indexOf('http') === 0) ? this.value : Admin.Url.upload(this.value))
         },
         link () {
             return ((this.file.indexOf('http') === 0) ? this.file : Admin.Settings.base_url + this.file)