+++ resources/assets/js/admin/form/image.js
@@ -56,17 +56,9 @@ Vue.component('element-image', Vue.extend({
         remove () {
             var self = this;
 
-            swal({
-                title: i18next.t('lang.message.are_you_sure'),
-                type: 'warning',
-                showCancelButton: true,
-                confirmButtonColor: '#3085d6',
-                cancelButtonColor: '#d33',
-                confirmButtonText: i18next.t('lang.button.yes')
-            }).then(() => {
+            Admin.Messages.confirm(trans('lang.message.are_you_sure')).then(() => {
                 self.value = '';
-            }, dismiss => {
-
+                self.file = '';
             });
         },
         closeAlert () {
@@ -78,6 +70,7 @@ Vue.component('element-image', Vue.extend({
             return this.value.length > 0
         },
         image () {
+            return ((this.value.indexOf('http') === 0) ? this.value : Admin.Url.upload(this.value))
             return ((this.file.indexOf('http') === 0) ? this.file : Admin.Settings.base_url + this.file)
         }
     }