+++ src/Display/DisplayTabbed.php
@@ -188,7 +188,9 @@ class DisplayTabbed implements DisplayInterface, FormInterface
     public function validateForm(\Illuminate\Http\Request $request, ModelConfigurationInterface $model = null)
     {
         $this->getTabs()->each(function ($tab) use ($request, $model) {
-            if ($tab instanceof FormInterface) {
+            $tabId = $request->get('sleeping_owl_tab_id');
+
+            if ($tab instanceof FormInterface && $tab->getName() == $tabId) {
                 $tab->validateForm($request, $model);
             }
         });
@@ -203,7 +205,9 @@ class DisplayTabbed implements DisplayInterface, FormInterface
     public function saveForm(\Illuminate\Http\Request $request, ModelConfigurationInterface $model = null)
     {
         $this->getTabs()->each(function (TabInterface $tab) use ($request, $model) {
-            if ($tab instanceof FormInterface) {
+            $tabId = $request->get('sleeping_owl_tab_id');
+
+            if ($tab instanceof FormInterface && $tab->getName() == $tabId) {
                 $tab->saveForm($request, $model);
             }
         });