+++ src/Display/Extension/ColumnFilters.php
@@ -135,11 +135,12 @@ class ColumnFilters extends Extension implements Initializable, Placable
      */
     public function toArray()
     {
+        $this->setHtmlAttribute('data-display', class_basename($this->getDisplay()));
+
         return [
             'filters' => $this->columnFilters,
             'attributes' => $this->htmlAttributesToString(),
             'tag' => $this->getPlacement() == 'table.header' ? 'thead' : 'tfoot',
-            'displayClass' => get_class($this->getDisplay()),
         ];
     }